sayt/api/v1/Category/serialazer.py
from rest_framework import serializers

from sayt_api.models import Category


class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = '__all__'
